Reorganize drivers definitions location
