+++ src/vs/workbench/contrib/positronConnections/browser/components/newConnectionModalDialog/listDriversState.tsx
@@ -11,96 +11,97 @@ import { LanguageRuntimeMetadata } from 'positron';
 import { DropDownListBox } from 'vs/workbench/browser/positronComponents/dropDownListBox/dropDownListBox';
 import { DropDownListBoxItem } from 'vs/workbench/browser/positronComponents/dropDownListBox/dropDownListBoxItem';
 import 'vs/css!./listDriversState';
-import { Driver, Input, InputType } from 'vs/workbench/contrib/positronConnections/browser/components/newConnectionModalDialog/driver';
+import { IDriver } from 'vs/workbench/services/positronConnections/browser/interfaces/positronConnectionsDriver';
 
 interface ListDriversProps {
 	readonly services: PositronConnectionsServices;
 	readonly onCancel: () => void;
-	readonly onSelection: (driver: Driver) => void;
+	readonly onSelection: (driver: IDriver) => void;
 }
 
 export const ListDrivers = (props: PropsWithChildren<ListDriversProps>) => {
 
-	const onDriverSelectedHandler = (driver: Driver) => {
+	const onDriverSelectedHandler = (driver: IDriver) => {
 		props.onSelection(driver);
 	};
 
 	const entries = getRegisteredLanguages(props.services);
 	const [languageId, setLanguageId] = useState<string>(entries[0]?.languageId);
-	const drivers = languageId ? getRegisteredDrivers(languageId) : [];
+
+	const driverManager = props.services.connectionsService.driverManager;
+
+	const drivers = languageId ?
+		driverManager.getDrivers().filter(driver => driver.languageId === languageId) :
+		[];
 
 	const onLanguageChangeHandler = (lang: string) => {
 		setLanguageId(lang);
 	};
 
 	return <div className='connections-new-connection-list-drivers'>
-		{
-			drivers.length === 0 ?
-				<div className='no-drivers'>
-					{localize('positron.newConnectionModalDialog.listDrivers.noDrivers', "No drivers available")}
-				</div> :
-				<>
-					<div className='title'>
-						<h1>
-							{localize('positron.newConnectionModalDialog.listDrivers.title', "Choose a Database Driver")}
-						</h1>
-					</div>
-					<div className='select-language'>
-						<DropDownListBox
-							keybindingService={props.services.keybindingService}
-							layoutService={props.services.layoutService}
-							title={localize('positron.newConnectionModalDialog.listDrivers.selectLanguage', "Select a language")}
-							entries={getRegisteredLanguages(props.services).map((item) => {
-								return new DropDownListBoxItem({
-									identifier: item.languageId,
-									value: item
-								});
-							})}
-							createItem={(item) => {
-								const value = item.options.value;
+		<div className='title'>
+			<h1>
+				{localize('positron.newConnectionModalDialog.listDrivers.title', "Choose a Database Driver")}
+			</h1>
+		</div>
+		<div className='select-language'>
+			<DropDownListBox
+				keybindingService={props.services.keybindingService}
+				layoutService={props.services.layoutService}
+				title={localize('positron.newConnectionModalDialog.listDrivers.selectLanguage', "Select a language")}
+				entries={getRegisteredLanguages(props.services).map((item) => {
+					return new DropDownListBoxItem({
+						identifier: item.languageId,
+						value: item
+					});
+				})}
+				createItem={(item) => {
+					const value = item.options.value;
 
-								return <div className='language-dropdown-entry'>
-									{value.base64EncodedIconSvg ? <img className='dropdown-entry-icon' src={`data:image/svg+xml;base64,${value.base64EncodedIconSvg}`} /> : null}
-									<div className='dropdown-entry-title'>
-										{value.languageName}
-									</div>
-								</div>;
-							}}
-							onSelectionChanged={(item) => onLanguageChangeHandler(item.options.identifier)}
-							selectedIdentifier={languageId}
-						>
-						</DropDownListBox>
-					</div>
-					<div className='driver-list'>
-						{
-							drivers.concat(drivers, drivers, drivers, drivers, drivers, drivers).map(driver => {
-								const icon = driver.base64EncodedIconSvg ?
-									<img className='driver-icon' src={`data:image/svg+xml;base64,${driver.base64EncodedIconSvg}`} /> :
-									<div className='driver-icon codicon codicon-database' style={{ opacity: 0.5, fontSize: '24px' }}></div>;
+					return <div className='language-dropdown-entry'>
+						{value.base64EncodedIconSvg ? <img className='dropdown-entry-icon' src={`data:image/svg+xml;base64,${value.base64EncodedIconSvg}`} /> : null}
+						<div className='dropdown-entry-title'>
+							{value.languageName}
+						</div>
+					</div>;
+				}}
+				onSelectionChanged={(item) => onLanguageChangeHandler(item.options.identifier)}
+				selectedIdentifier={languageId}
+			>
+			</DropDownListBox>
+		</div>
+		<div className='driver-list'>
+			{
+				drivers.length > 0 ?
+					drivers.map(driver => {
+						const icon = driver.base64EncodedIconSvg ?
+							<img className='driver-icon' src={`data:image/svg+xml;base64,${driver.base64EncodedIconSvg}`} /> :
+							<div className='driver-icon codicon codicon-database' style={{ opacity: 0.5, fontSize: '24px' }}></div>;
 
-								return <div key={driver.driverId} className='driver-list-item'>
-									{icon}
-									<div className='driver-info' onMouseDown={() => onDriverSelectedHandler(driver)}>
-										<div className='driver-name'>
-											{driver.name}
-										</div>
-										<div className={`driver-button codicon codicon-chevron-right`}>
-										</div>
-									</div>
-								</div>;
-							})
-						}
+						return <div key={driver.driverId} className='driver-list-item'>
+							{icon}
+							<div className='driver-info' onMouseDown={() => onDriverSelectedHandler(driver)}>
+								<div className='driver-name'>
+									{driver.name}
+								</div>
+								<div className={`driver-button codicon codicon-chevron-right`}>
+								</div>
+							</div>
+						</div>;
+					}) :
+					<div className='no-drivers'>
+						{localize('positron.newConnectionModalDialog.listDrivers.noDrivers', "No drivers available")}
 					</div>
-					<div className='footer'>
-						<PositronButton
-							className='button action-bar-button'
-							onPressed={props.onCancel}
-						>
-							{(() => localize('positron.resumeConnectionModalDialog.cancel', "Cancel"))()}
-						</PositronButton>
-					</div>
-				</>
-		}
+			}
+		</div>
+		<div className='footer'>
+			<PositronButton
+				className='button action-bar-button'
+				onPressed={props.onCancel}
+			>
+				{(() => localize('positron.resumeConnectionModalDialog.cancel', "Cancel"))()}
+			</PositronButton>
+		</div>
 	</div>;
 };
 
@@ -114,83 +115,4 @@ const getRegisteredLanguages = (services: PositronConnectionsServices) => {
 		languages.set(runtime.languageId, preferedMetadata);
 	}
 	return Array.from(languages.values());
-};
-
-const getRegisteredDrivers = (languageId: string): Array<Driver> => {
-	// TODO currently we always return the same list of drivers.
-	// but we we'll have a mechanism for extensions to register drivers
-	// for a given language.
-	return [
-		{
-			languageId: languageId,
-			driverId: 'postgres',
-			name: 'PostgresSQL',
-			inputs: [
-				{
-					'id': 'dbname',
-					'label': 'Database Name',
-					'type': InputType.String,
-					'value': 'localhost'
-				},
-				{
-					'id': 'host',
-					'label': 'Host',
-					'type': InputType.String,
-					'value': 'localhost'
-				},
-				{
-					'id': 'port',
-					'label': 'Port',
-					'type': InputType.Number,
-					'value': '5432'
-				},
-				{
-					'id': 'user',
-					'label': 'User',
-					'type': InputType.String,
-					'value': 'postgres'
-				},
-				{
-					'id': 'password',
-					'label': 'Password',
-					'type': InputType.String,
-					'value': 'password'
-				},
-				{
-					'id': 'bigint',
-					'label': 'Integer representation',
-					'type': InputType.Option,
-					'options': [
-						{ 'identifier': 'integer64', 'title': 'integer64' },
-						{ 'identifier': 'integer', 'title': 'integer' },
-						{ 'identifier': 'numeric', 'title': 'numeric' },
-						{ 'identifier': 'character', 'title': 'character' }
-					],
-					'value': 'integer64'
-				}
-			],
-			generateCode: (inputs: Array<Input>) => {
-				const dbname = inputs.find(input => input.id === 'dbname')?.value;
-				const host = inputs.find(input => input.id === 'host')?.value;
-				const port = inputs.find(input => input.id === 'port')?.value;
-				const user = inputs.find(input => input.id === 'user')?.value;
-				const password = inputs.find(input => input.id === 'password')?.value;
-				const bigint = inputs.find(input => input.id === 'bigint')?.value;
-
-
-				return `library(DBI)
-con <- dbConnect(
-	RPostgres::Postgres(),
-	dbname = '${dbname ?? ''}',
-	host = '${host ?? ''}',
-	port = ${port ?? ''},
-	user = '${user ?? ''}',
-	password = '${password ?? ''}',
-	bigint = '${bigint ?? ''}'
-)
-
-`;
-			}
-		},
-	];
 };
